TransportationMatrix
Have several choices for the transportation type. There may not a direct path from you the source to the destination city, and thus you have to pass some other cities to reach the destination. Passing through the cities, you need to choose the transportation type of the path connecting  these cities.